+++ CHANGELOG.md
@@ -141,6 +141,7 @@
         * Update package name to `com.braintreepayments.api.sharedutils`
     * PayPal Messaging (BETA) 
         * Remove `BraintreeClient` from constructor
+        * Update package name to `com.braintreepayments.api.paypalmessaging`
     * Shopper Insights (BETA)
         * Remove `BraintreeClient` from constructor
