+++ PayPalMessaging/src/main/java/com/braintreepayments/api/paypalmessaging/PayPalMessagingAnalytics.kt
@@ -1,4 +1,4 @@
-package com.braintreepayments.api
+package com.braintreepayments.api.paypalmessaging
 
 internal object PayPalMessagingAnalytics {
     const val STARTED = "paypal-messaging:create-view:started"
