+++ PayPalMessaging/src/main/java/com/braintreepayments/api/paypalmessaging/PayPalMessagingView.kt
@@ -1,11 +1,11 @@
-package com.braintreepayments.api
+package com.braintreepayments.api.paypalmessaging
 
 import android.content.Context
 import android.view.ViewGroup
 import android.widget.FrameLayout
+import com.braintreepayments.api.ExperimentalBetaApi
 import com.braintreepayments.api.core.BraintreeClient
 import com.braintreepayments.api.core.BraintreeException
-import com.braintreepayments.api.core.BuildConfig
 import com.braintreepayments.api.core.Configuration
 import com.paypal.messages.PayPalMessageView
 import com.paypal.messages.config.PayPalEnvironment
