+++ PayPalMessaging/src/test/java/com/braintreepayments/api/paypalmessaging/PayPalMessagingLogoTypeUnitTest.kt
@@ -1,5 +1,6 @@
-package com.braintreepayments.api
+package com.braintreepayments.api.paypalmessaging
 
+import com.braintreepayments.api.ExperimentalBetaApi
 import com.paypal.messages.config.message.style.PayPalMessageLogoType
 import junit.framework.TestCase.assertEquals
 import org.junit.Test
